refactor(jupiter): rename `TransactionProcessor` to `TransactionPreProcessor`
