+++ gravitee-apim-gateway/gravitee-apim-gateway-reactor/src/main/java/io/gravitee/gateway/jupiter/reactor/processor/transaction/TransactionProcessorFactory.java
@@ -15,15 +15,15 @@
  */
 package io.gravitee.gateway.jupiter.reactor.processor.transaction;
 
+import static io.gravitee.gateway.jupiter.reactor.processor.transaction.TransactionHeader.DEFAULT_REQUEST_ID_HEADER;
+import static io.gravitee.gateway.jupiter.reactor.processor.transaction.TransactionHeader.DEFAULT_TRANSACTION_ID_HEADER;
+
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
  * @author Guillaume LAMIRAND (guillaume.lamirand at graviteesource.com)
  * @author GraviteeSource Team
  */
 public class TransactionProcessorFactory {
-
-    public static final String DEFAULT_TRANSACTION_ID_HEADER = "X-Gravitee-Transaction-Id";
-    public static final String DEFAULT_REQUEST_ID_HEADER = "X-Gravitee-Request-Id";
 
     private final String transactionHeader;
     private final String requestHeader;
@@ -33,7 +33,7 @@ public class TransactionProcessorFactory {
         this.requestHeader = requestHeader == null ? DEFAULT_REQUEST_ID_HEADER : requestHeader;
     }
 
-    public TransactionProcessor create() {
-        return new TransactionProcessor(transactionHeader, requestHeader);
+    public TransactionPreProcessor create() {
+        return new TransactionPreProcessor(transactionHeader, requestHeader);
     }
 }
