+++ gravitee-apim-gateway/gravitee-apim-gateway-reactor/src/test/java/io/gravitee/gateway/jupiter/reactor/processor/transaction/TransactionProcessorFactoryTest.java
@@ -15,8 +15,8 @@
  */
 package io.gravitee.gateway.jupiter.reactor.processor.transaction;
 
-import static io.gravitee.gateway.jupiter.reactor.processor.transaction.TransactionProcessorFactory.DEFAULT_REQUEST_ID_HEADER;
-import static io.gravitee.gateway.jupiter.reactor.processor.transaction.TransactionProcessorFactory.DEFAULT_TRANSACTION_ID_HEADER;
+import static io.gravitee.gateway.jupiter.reactor.processor.transaction.TransactionHeader.DEFAULT_REQUEST_ID_HEADER;
+import static io.gravitee.gateway.jupiter.reactor.processor.transaction.TransactionHeader.DEFAULT_TRANSACTION_ID_HEADER;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.Test;
@@ -33,7 +33,7 @@ public class TransactionProcessorFactoryTest {
             DEFAULT_TRANSACTION_ID_HEADER,
             DEFAULT_REQUEST_ID_HEADER
         );
-        TransactionProcessor transactionProcessor = transactionProcessorFactory.create();
+        TransactionPreProcessor transactionProcessor = transactionProcessorFactory.create();
         assertThat(transactionProcessor.transactionHeader()).isEqualTo(DEFAULT_TRANSACTION_ID_HEADER);
         assertThat(transactionProcessor.requestHeader()).isEqualTo(DEFAULT_REQUEST_ID_HEADER);
     }
@@ -44,7 +44,7 @@ public class TransactionProcessorFactoryTest {
             "CUSTOM_TRANSACTION_ID_HEADER",
             "CUSTOM_REQUEST_ID_HEADER"
         );
-        TransactionProcessor transactionProcessor = transactionProcessorFactory.create();
+        TransactionPreProcessor transactionProcessor = transactionProcessorFactory.create();
         assertThat(transactionProcessor.transactionHeader()).isEqualTo("CUSTOM_TRANSACTION_ID_HEADER");
         assertThat(transactionProcessor.requestHeader()).isEqualTo("CUSTOM_REQUEST_ID_HEADER");
     }
